Feature/1133 fix deprecated refs (#13)
* Move order questions and service class to product

* Fix indentation for Rubocop

* Create products from product class in demo
