+++ app/models/jellyfish_demo/product_type/widget.rb
@@ -23,17 +23,11 @@ module JellyfishDemo
 
       def product_questions
         [
-          { name: :environment, value_type: :string, field: :environments, required: true }
         ]
       end
 
-      def order_questions
-        [
-        ]
-      end
-
-      def service_class
-        'JellyfishDemo::Service::Widget'.constantize
+      def product_class
+        'JellyfishDemo::Product::Widget'.constantize
       end
     end
   end
